refactor: move cli UI primitives into cli/ui/ subpackage
Move _ansi, _terminal, _timed_input, _menu to cli/ui/ and update all
importers across cli/, session/, fleet/, and flat modules.
