+++ src/autoskillit/cli/fleet/_fleet_lifecycle.py
@@ -264,7 +264,7 @@ def _reap_stale_dispatches(state_path: Path, *, dry_run: bool = False) -> None:
 
 def _pick_resume_campaign(project_dir: Path) -> tuple[str, str]:
     """Interactively pick a resumable campaign. Returns (campaign_name, campaign_id) or exits."""
-    from autoskillit.cli._menu import run_selection_menu  # noqa: PLC0415
+    from autoskillit.cli.ui._menu import run_selection_menu  # noqa: PLC0415
     from autoskillit.fleet import TERMINAL_DISPATCH_STATUSES, read_state  # noqa: PLC0415
 
     fleet_dir = project_dir / ".autoskillit" / "temp" / "fleet"
