+++ src/autoskillit/cli/ui/__init__.py
@@ -1,0 +1,1 @@
+"""CLI UI primitives — color, terminal, prompts, menus."""
