+++ src/autoskillit/cli/ui/_timed_input.py
@@ -14,7 +14,7 @@ from __future__ import annotations
 import select
 import sys
 
-from autoskillit.cli._ansi import supports_color
+from autoskillit.cli.ui._ansi import supports_color
 from autoskillit.cli._init_helpers import _require_interactive_stdin
